Fixed API communication (calendar)
