+++ credentials_calendar.json
@@ -1,1 +1,6 @@
-{"web":{"client_id":"623975066925-khsier4amnf5t3ndkfok4mgvf51ugtb8.apps.googleusercontent.com","project_id":"skilled-text-315319","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost:50635/"]}}
+{"web":{"client_id":"623975066925-khsier4amnf5t3ndkfok4mgvf51ugtb8.apps.googleusercontent.com",
+    "project_id":"skilled-text-315319",
+    "auth_uri":"https://accounts.google.com/o/oauth2/auth",
+    "token_uri":"https://accounts.google.com/o/oauth2/token",
+    "client_secret":"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",
+    "redirect_uris":["http://localhost:50635/"]}}
