+++ token-calendar.json
@@ -1,0 +1,1 @@
+{"token": "ya29.a0ARrdaM95EwzkxaEYjVY6VMPkraeT2HgCN96pqV47bTFZtEvMGDuXbHdf6QiDWSFblRZ2UItixHNTebmXaYLdq2dDjTmqrAAuzjIqFW0gWCgQC0QMTeFNNIRbQyxRq-7zL2SvO2uCLMU0Fs1hY08EwyCkiQad", "refresh_token": "1//0f1v4FF7JQSs1CgYIARAAGA8SNwF-L9IrgUluS_42ann1P62CZwW3P182AvytD6QnfVSbG68eY0KAMFDlZ_BypE06khNO7ZhqGs0", "token_uri": "https://accounts.google.com/o/oauth2/token", "client_id": "623975066925-khsier4amnf5t3ndkfok4mgvf51ugtb8.apps.googleusercontent.com", "client_secret": "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", "scopes": ["https://www.googleapis.com/auth/calendar.readonly"], "expiry": "2022-05-16T03:16:51.645035Z"}
