+++ wsgi.py
@@ -1,4 +1,4 @@
-from main import create_app
+from run import app
 
 if __name__ == "__main__":
   app.run()
